requirements.txt
# python 3.7.5
# pip 23.3.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==1.5.4
certifi==2020.12.5
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
constantly==15.1.0
cryptography==42.0.2
daphne==4.0.0
Django==3.2.23
django-cors-headers==3.7.0
django-redis==4.12.1
djangorestframework==3.12.2
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.7.0
incremental==22.10.0
msgpack==1.0.5
mysqlclient==2.0.3
Pillow==8.1.2
pkg_resources==0.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.6.0
pycparser==2.21
pyOpenSSL==24.0.0
pytz==2020.5
redis==5.0.1
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==1.26.3
zipp==3.15.0
zope.interface==6.1